my project/requirements.txt
flask
faiss-cpu
sentence-transformers
pytesseract
Pillow
openai
beautifulsoup4
tqdm
requests